+++ src/interface/common.ts
@@ -1,2 +1,2 @@
-export type Formatter<T, P> = (payload?: P) => T;
+export type Formatter<T, P> = (payload: P) => T;
 export type BaseAttr<T, P> = T | Formatter<T, P>;